Isolate impersonation feature to demo-only build target
Move all impersonation code (endpoints, model, auth token handling) into
app/demo/ and app/demo_main.py, which are present only in the demo
container image (Dockerfile.demo). The production Dockerfile explicitly
removes app/demo/ after COPY to make the exclusion auditable.

Production app.main has zero knowledge of impersonation. The demo entry
point (app.demo_main) extends the prod app at startup by:
- Importing demo models so init_db() creates the impersonation_sessions table
- Overriding the get_current_user dependency with an impersonation-aware version
- Registering the impersonation API router
