+++ Dockerfile.demo
@@ -1,60 +1,53 @@
-# Multi-stage build for GHARTS backend
+# Demo image — includes impersonation endpoints. NOT for production use.
+#
+# Build:  docker build -f Dockerfile.demo -t gha-runner-token-service:demo .
+# Run:    docker run -e DEMO_IMPERSONATION_SECRET=<secret> ... gha-runner-token-service:demo
+
 # Stage 1: Builder - Install dependencies
 FROM python:3.11-slim AS builder
 
 WORKDIR /build
 
-# Install build dependencies
 RUN apt-get update && apt-get install -y --no-install-recommends \
     gcc \
     libpq-dev \
     && rm -rf /var/lib/apt/lists/*
 
-# Copy requirements and install Python dependencies
 COPY requirements.txt .
 RUN pip install --no-cache-dir --user -r requirements.txt
 
-# Stage 2: Runtime - Minimal production image
+# Stage 2: Runtime
 FROM python:3.11-slim AS runtime
 
-# Create non-root user
 RUN groupadd -r appuser -g 1000 && \
     useradd -r -u 1000 -g appuser -m -s /sbin/nologin appuser
 
 WORKDIR /app
 
-# Install runtime dependencies only
 RUN apt-get update && apt-get install -y --no-install-recommends \
     libpq5 \
     curl \
     && rm -rf /var/lib/apt/lists/*
 
-# Copy Python dependencies from builder
 COPY --from=builder /root/.local /home/appuser/.local
 
-# Copy application code — demo modules are intentionally excluded
+# Copy full app including demo modules
 COPY --chown=appuser:appuser app/ ./app/
-RUN rm -rf /app/app/demo /app/app/demo_main.py
 
-# Create necessary directories with proper permissions
 RUN mkdir -p /app/logs /app/data /tmp/app && \
     chown -R appuser:appuser /app/logs /app/data /tmp/app
 
-# Set environment variables
 ENV PATH=/home/appuser/.local/bin:$PATH \
     PYTHONUNBUFFERED=1 \
     PYTHONDONTWRITEBYTECODE=1 \
     LOG_DIR=/app/logs
 
-# Switch to non-root user
 USER appuser
 
-# Expose port
 EXPOSE 8000
 
-# Health check
 HEALTHCHECK --interval=30s --timeout=3s --start-period=40s --retries=3 \
     CMD curl -f http://localhost:8000/health || exit 1
 
-# Run application
-CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8000"]
+# Use demo_main instead of main — this is the only difference from the prod image
+CMD ["uvicorn", "app.demo_main:app", "--host", "0.0.0.0", "--port", "8000"]
