+++ app/demo/__init__.py
@@ -1,0 +1,1 @@
+"""Demo-only features. Not included in production container images."""
